read_textfile: function that reads a file and prints their content in stdout, using long

+++ 0x15-file_io/0-read_textfile.c
@@ -11,7 +11,7 @@
  * @letters: number of chars to be read and printed
  * Return: the number of chars printed or 0 if error
  */
-long read_textfile(const char *filename, size_t letters)
+long read_textfile(const char *filename, long letters)
 {
 	int fd;
 	long read_chars, write_chars;
